fixed some compiler warnings
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@2345 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ sc_rng.cpp
@@ -364,7 +364,7 @@ struct normalized_distance_rng_t : public normalized_rng_t
   void setN4(double expP)
   {
     lastAvg = expP;
-    N1 = 1 / expP;
+    N1 = (int) (1 / expP);
     if (N1 > maxN4) N1 = maxN4;
     N4 = 4 * N1;
     if (N4 > maxN4) N4 = maxN4;
